Extract unmaterializable functions to own file
Extract the unmaterializable functions to their own file, in order to
make `func.rs` smaller.
